Move the field types into traits
