+++ admin/AdminPage/FixesPage.php
@@ -16,6 +16,9 @@ namespace EqualizeDigital\AccessibilityChecker\Admin\AdminPage;
  */
 class FixesPage implements PageInterface {
 
+	use FixesSettingType\Checkbox;
+	use FixesSettingType\Text;
+
 	/**
 	 * The capability required to access the settings page.
 	 *
@@ -177,42 +180,5 @@ class FixesPage implements PageInterface {
 	 */
 	public function fixes_section_general_cb() {
 		echo '<p>' . esc_html__( 'General settings for the fixes.', 'accessibility-checker' ) . '</p>';
-	}
-
-	/**
-	 * Render a checkbox input.
-	 *
-	 * @param array $args The arguments for the checkbox. This is expected to have a name and a description.
-	 */
-	public static function checkbox( $args ) {
-
-		// We need a name and a description or the checkbox is useless.
-		if ( ! isset( $args['name'], $args['description'] ) ) {
-			return;
-		}
-
-		$option_value = get_option( $args['name'] );
-		?>
-		<label>
-			<input
-				type="checkbox"
-				value="1"
-				id="<?php echo esc_attr( $args['name'] ); ?>"
-				name="<?php echo esc_attr( $args['name'] ); ?>"
-				<?php checked( 1, $option_value ); ?>
-			/>
-			<?php echo esc_html( $args['description'] ); ?>
-		</label>
-		<?php
-	}
-
-	/**
-	 * Sanitize a checkbox input.
-	 *
-	 * @param mixed $input The input to sanitize.
-	 * @return int
-	 */
-	public function sanitize_checkbox( $input ) {
-		return isset( $input ) ? 1 : 0;
 	}
 }
